changes massless parts processing behavior
In KSP the mass of massless parts (parts with attribute
PhysicsSignificance = 1 ,
http://wiki.kerbalspaceprogram.com/wiki/Massless_part) is not taken into
account, although stated in the description.
However, the original Ubiozur code adds the weight of such parts to the
total mass. of welded part.
This patch allows you to make the mass processing similar to stock, if
settings file config.xml has line
<bool name="DontProcessMasslessParts">1</bool>
Any other attributes processed as usual.